Add cypress function for expecting a certain documentStatus + helper class
https://github.com/metasfresh/metasfresh-e2e/issues/133
